가격이 숫자가 아니라서 숫자로 선언하도록 고침

+++ modules/nmileage/nmileage.model.php
@@ -91,7 +91,7 @@ class nmileageModel extends nmileage
 
 	function getMileageByProduct()
 	{
-		$price = Context::get('price');
+		$price = (int)Context::get('price');
 
 		$oNcartModel = getModel('ncart');
 		$config = $oNcartModel->getModuleConfig();
